chore: delete unused downloaders for ibge open data portal and SIDRA and refactor localidades.py
