--- datatools/downloaders/ibge-dados-abertos/ibge/README.md
@@ -1,5 +1,0 @@
-# Execução
-
-```
-scrapy crawl ibge
-```
