--- datatools/downloaders/ibge-dados-abertos/ibge/ibge/items.py
@@ -1,12 +1,0 @@
-# Define here the models for your scraped items
-#
-# See documentation in:
-# https://docs.scrapy.org/en/latest/topics/items.html
-
-import scrapy
-
-
-class IbgeItem(scrapy.Item):
-    # define the fields for your item here like:
-    file_urls = scrapy.Field()
-    files = scrapy.Field()
