--- datatools/downloaders/ibge-dados-abertos/requirements.txt
@@ -1,2 +1,0 @@
-itemadapter==0.8.0
-Scrapy==2.11.0
